[test | refactor] mock데이터가 든 폴더명 변경
- __test__에서 __mocks__로 변경
