+++ src/components/UI/organism/SkillList/ProfileSkillCard.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { withKnobs, text } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import SkillList from '.';
-import { skills } from '../../../../__test__/mock-dats';
+import { skills } from '../../../../__mocks__/mock-dats';
 
 export default {
   title: 'Organism/SkillList',
