+++ src/pages/Signup/Signup.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import MockAdapter from 'axios-mock-adapter';
 import SignupTemplate from './template';
 import AxiosMock from '../../common/AxiosMock';
-import { signupSuccessMockInfo, signupFailMockInfo } from '../../__test__/mock-dats';
+import { signupSuccessMockInfo, signupFailMockInfo } from '../../__mocks__/mock-dats';
 
 export default {
   title: 'Template/Signup',
