+++ src/pages/__test__/Signup.test.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { fireEvent, waitFor } from '@testing-library/react';
 import { render } from '../../test-utils';
 import SignupPage from '../Signup';
-import { signupFailMockInfo } from '../../__test__/mock-dats';
+import { signupFailMockInfo } from '../../__mocks__/mock-dats';
 import { signupAPI } from '../../api/auth';
 
 jest.mock('../../api/auth');
